+++ packages/expo-video/ios/VideoManager.swift
@@ -10,15 +10,38 @@ import ExpoModulesCore
 class VideoManager {
   static var shared = VideoManager()
 
+  private var managerQueue = DispatchQueue(label: "com.expo.video.manager.managerQueue")
   private var videoViews = NSHashTable<VideoView>.weakObjects()
   private var videoPlayers = NSHashTable<VideoPlayer>.weakObjects()
 
-  func register(videoPlayer: VideoPlayer) {
-    videoPlayers.add(videoPlayer)
+  func register(videoPlayer: VideoPlayer, queue: DispatchQueue? = nil) {
+    let queue = queue ?? managerQueue
+    queue.async { [weak self, weak videoPlayer] in
+      guard let self = self, let videoPlayer = videoPlayer else {
+        return
+      }
+      self.videoPlayers.add(videoPlayer)
+    }
   }
 
-  func unregister(videoPlayer: VideoPlayer) {
-    videoPlayers.remove(videoPlayer)
+  func register(videoPlayer: VideoPlayer, queue: DispatchQueue? = nil) async {
+    let queue = queue ?? managerQueue
+    await queue.async { [weak self, weak videoPlayer] in
+      guard let self = self, let videoPlayer = videoPlayer else {
+        return
+      }
+      self.videoPlayers.add(videoPlayer)
+    }
+  }
+
+  func unregister(videoPlayer: VideoPlayer, queue: DispatchQueue? = nil) {
+    let queue = queue ?? managerQueue
+    queue.async { [weak self, weak videoPlayer] in
+      guard let self = self, let videoPlayer = videoPlayer else {
+        return
+      }
+      self.videoPlayers.remove(videoPlayer)
+    }
   }
 
   func register(videoView: VideoView) {
@@ -50,7 +73,16 @@ class VideoManager {
 
   // MARK: - Audio Session Management
 
-  internal func setAppropriateAudioSessionOrWarn() {
+  // This function usually takes less than 5ms to execute, but in some cases (initial setup) it takes up to 70ms
+  // Because of this we dispatch it on another queue to minimize the load on main queue.
+  internal func setAppropriateAudioSessionOrWarn(queue: DispatchQueue? = nil) {
+    let queue = queue ?? managerQueue
+    queue.async { [weak self] in
+      self?.setAudioSession()
+    }
+  }
+
+  private func setAudioSession() {
     let audioSession = AVAudioSession.sharedInstance()
     let audioMixingMode = findAudioMixingMode()
     var audioSessionCategoryOptions: AVAudioSession.CategoryOptions = audioSession.categoryOptions